referral/requirements.txt
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
fastapi==0.54.2
h11==0.9.0
httptools==0.1.1
idna==2.9
lorem==0.1.1
mysql-connector-python==8.0.20
netifaces==0.10.9
protobuf==3.12.0
pycparser==2.20
pydantic==1.5.1
PyJWT==1.7.1
python-consul==1.1.0
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.16
starlette==0.13.2
urllib3==1.25.9
uvicorn==0.11.5
uvloop==0.14.0
websockets==8.1
